Get rid of leaf struct and use the existing node struct for leaves. This adds two (unneeded) pointers to leaves, but makes the code simpler and less suspectible to error.
Leaf/non-leaf-ness is signaled through the node color anyway;
the leaf struct was not being used for anything distinct.

Adds about 1MB of space per 32,768 keys*. Unless someone is storing
Oxford's English Dictionary on a wristwatch, this should be fine.

  * Calculation: 1024^2 = 2x ( 8 * 2 ), solve for x

    1024^2 = bytes in a megabyte
    2x = nodes in a completely balanced tree for x keys
    8 * 2 = two pointers, each 8 bytes each (on 64-bit).

On the upside, a 32-bit wristwatch would be able to store twice
as many keys.