Small but important change: make the keys consistent down the tree. Avoid 'orphaned' keys; that is, each inner node key traces a path down to its leaf.
This will be important when properly resetting inner node keys
after deletion, as it avoids having to search up the entire tree
for orphans belonging to the deleted leaf.

Also, prefer older keys for inner nodes. This further enforces
the guarantee of an inner key making a path to its leaf.